[GLUEGEN] create parse_helpers.cpp
containing all loose functions (simple strtok, braces parsing..)
